requirements.txt

requests
scikit-learn
pypdf
langchain
langchain-google-genai
langchain-community
langchain-huggingface
faiss-cpu
chromadb
Pillow
pytesseract
sentence-transformers
langchain-groq
python-dotenv
fastapi
uvicorn
python-multipart
pdf2image